// frontend/src/config/enums.ts
export enum Locale {
    EN = 'english',
    IT = 'italian',
}

export enum RefreshInterval {
    OFF = 'OFF',
    FIVE_SECONDS = 5000,
    TEN_SECONDS = 10000,
    THIRTY_SECONDS = 30000,
    ONE_MINUTE = 60000,
    TWO_MINUTES = 120000,
    FIVE_MINUTES = 300000,
    FIFTEEN_MINUTES = 900000,
    THIRTY_MINUTES = 1800000,
    ONE_HOUR = 3600000,
    TWO_HOURS = 7200000,
    ONE_DAY = 86400000,
}